requirements.txt
flask
pillow
torch
transformers
